Split library source and utility source
