+++ seekgzip.c
@@ -22,7 +22,7 @@
 
 #define SEEKGZIP_OPTIMIZATION
 
-/*===== Begin of the portion of zran.c =====*/
+/*===== Begin of the portion of zran.c ===== {{{*/
 
 /* zran.c -- example of zlib/gzip stream indexing and random access
  * Copyright (C) 2005 Mark Adler
@@ -393,9 +393,14 @@ static int extract(FILE *in, struct access *index, off_t offset,
     return ret;
 }
 
-/*===== End of the portion of zran.c =====*/
+/*===== End of the portion of zran.c ===== }}}*/
 
-
+struct tag_seekgzip {
+    FILE                      *fp;
+    struct access              index;
+    off_t                      offset;
+    int                        errorcode;
+};
 
 static char *get_index_file(const char *target)
 {
@@ -419,14 +424,6 @@ static uint32_t read_uint32(gzFile gz)
     gzread(gz, &v, sizeof(v));
     return v;
 }
-
-struct tag_seekgzip
-{
-    FILE *fp;
-    struct access index;
-    off_t offset;
-    int errorcode;
-};
 
 int seekgzip_build(const char *target)
 {
@@ -652,115 +649,3 @@ int seekgzip_error(seekgzip_t* sgz)
 {
     return sgz->errorcode;
 }
-
-#ifdef BUILD_UTILITY
-
-static void seekgzip_perror(int ret)
-{
-    switch (ret) {
-    case SEEKGZIP_ERROR:
-        fprintf(stderr, "ERROR: An unknown error occurred.\n");
-        break;
-    case SEEKGZIP_OPENERROR:
-        fprintf(stderr, "ERROR: Failed to open a file.\n");
-        break;
-    case SEEKGZIP_READERROR:
-        fprintf(stderr, "ERROR: Failed to read a file.\n");
-        break;
-    case SEEKGZIP_WRITEERROR:
-        fprintf(stderr, "ERROR: Failed to write a file.\n");
-        break;
-    case SEEKGZIP_DATAERROR:
-        fprintf(stderr, "ERROR: The file is corrupted.\n");
-        break;
-    case SEEKGZIP_OUTOFMEMORY:
-        fprintf(stderr, "ERROR: Out of memory.\n");
-        break;
-    case SEEKGZIP_IMCOMPATIBLE:
-        fprintf(stderr, "ERROR: The imcompatible file.\n");
-        break;
-    case SEEKGZIP_ZLIBERROR:
-        fprintf(stderr, "ERROR: An error occurred in zlib.\n");
-        break;
-    }
-}
-
-int main(int argc, char *argv[])
-{
-    int ret = 0;
-
-    if (argc != 3) {
-        printf("This utility manages an index for random (seekable) access to a gzip file.\n");
-        printf("USAGE:\n");
-        printf("    %s -b <FILE>\n", argv[0]);
-        printf("        Build an index file \"$FILE.idx\" for the gzip file $FILE.\n");
-        printf("    %s <FILE> [BEGIN-END]\n", argv[0]);
-        printf("        Output the content of the gzip file $FILE of offset range [BEGIN:END).\n");
-        return 0;
-
-    } else if (strcmp(argv[1], "-b") == 0) {
-        const char *target = argv[2];
-
-        printf("Building an index: %s.idx\n", target);
-        printf("Filesize up to: %d bit\n", (int)sizeof(off_t) * 8);
-
-        ret = seekgzip_build(target);
-        if (ret != 0) {
-            seekgzip_perror(ret);
-            return 1;
-        }
-        return 0;
-
-    } else {
-        char *arg = argv[2], *p = NULL;
-        off_t begin = 0, end = (off_t)-1;
-        seekgzip_t* zs = seekgzip_open(argv[1], NULL);
-        if (zs == NULL) {
-            fprintf(stderr, "ERROR: Failed to open the index file.\n");
-            return 1;
-        }
-
-        p = strchr(arg, '-');
-        if (p == NULL) {
-            begin =(off_t)strtoull(arg, NULL, 10);
-            end = begin+1;
-        } else if (p == arg) {
-            begin = 0;
-            end = (off_t)strtoull(p+1, NULL, 10);
-        } else if (p == arg + strlen(arg) - 1) {
-            *p = 0;
-            begin = (off_t)strtoull(arg, NULL, 10);
-        } else {
-            *p++ = 0;
-            begin =(off_t)strtoull(arg, NULL, 10);
-            end =(off_t)strtoull(p, NULL, 10);
-        }
-
-        seekgzip_seek(zs, begin);
-
-        while (begin < end) {
-            int read;
-            char buffer[CHUNK];
-            off_t size = (end - begin);
-            if (CHUNK < size) {
-                size = CHUNK;
-            }
-            read = seekgzip_read(zs, buffer, (int)size);
-            if (0 < read) {
-                fwrite(buffer, read, sizeof(char), stdout);
-                begin += read;
-            } else if (read == 0) {
-                break;
-            } else {
-                fprintf(stderr, "ERROR: An error occurred while reading the gzip file.\n");
-                ret = 1;
-                break;
-            }
-        }
-    
-        seekgzip_close(zs);
-        return ret;
-    }
-}
-
-#endif
